update principles
help me understand what I am missing here... so we say only one writer at a time but we can mitigate this problem to some extent by redefinig the boundary of where the writer can write, right? for example, lets say we have an ecommerce company called rainforest inc. it has hundreds of thousands of products, hundreds of copy editors around the globe who need to write  / edit descriptions for these products but we can say something like when writer A edits the copy for product B, only the copy for product B is locked, nothing else, right? doesn't that help us mitigate the reader writer problem? or is it still a problem? please rewrite the answer above with an eye to this explanation and do not hesitate to over explain, even though I know we need to be concise. use real world examples where possible and lots of them! also remember to output the whole thing as a single markdown so it is easy for me to copy paste. be clear and concise but also comprehensive